refactor: change naming components
