+++ src/components/Layout/index.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom'
 
 import { Footer } from '../Footer'
-import { Header } from '../header/Header'
+import { Header } from '../Header'
 
 export const Layout = () => {
   return (
